requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
sqlalchemy==2.0.15
pymysql==1.0.3
pydantic==1.10.7
bcrypt==4.0.1
